refactor(core): Extract pagination helpers into dedicated module (to simplify request-helper-function.ts, preparation for more SSRF improvements)
