+++ packages/core/src/execution-engine/node-execution-context/utils/request-helpers/index.ts
@@ -12,3 +12,8 @@ export {
 	getUrlFromProxyConfig,
 	setAxiosAgents,
 } from './axios-utils';
+export {
+	applyPaginationRequestData,
+	requestWithAuthenticationPaginated,
+	type ResolveValueFn,
+} from './pagination';
